+++ src/common/js/checks.js
@@ -5,8 +5,6 @@
 import _ from 'lodash';
 const assert = require( 'assert' ).strict;
 
-import { DateJs } from 'meteor/pwix:date';
-import { pwixI18n } from 'meteor/pwix:i18n';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { TM } from 'meteor/pwix:typed-message';