+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:validity',
-    version: '1.1.0',
+    version: '1.1.1-rc',
     summary: 'A Meteor package to manage validities',
     git: 'https://github.com/trychlos/pwix-validity.git',
     documentation: 'README.md'