+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 ## ChangeLog
 
+### 1.1.1-rc
+
+    Release date: 
+
+    - 
+
 ### 1.1.0
 
     Release date: 2024- 7-18